!241 merge feature/official-hubs-spharx01 into feature/official-hubs

Created-by: spharx01
Commit-by: spharx01;zhixianzhou
Merged-by: spharx01
Description: # Pull Request

## 变更类型

- [ ] Bug 修复
- [ ] 新功能
- [ ] 重构 / 代码清理
- [ ] 性能优化
- [ ] 文档更新
- [ ] 测试补充

## 相关 Issue

Closes #____ (如有)

## 变更摘要

<!-- 一两句话说明改了什么、为什么改 -->

## 检查清单

提交前请确认以下事项：

### 构建 & 测试

- [ ] 本地构建通过：
  ```bash
  cmake -B ../AgentRT-build -DBUILD_TESTS=ON -DCMAKE_BUILD_TYPE=Release
  cmake --build ../AgentRT-build --parallel 8
  ```
- [ ] 测试通过：`cd ../AgentRT-build && ctest --output-on-failure`
- [ ] 代码风格通过：`clang-format --dry-run --Werror` 对所有变更的 `.c`/`.h` 文件

### 代码质量

- [ ] Commit 遵循 [Conventional Commits](https://www.conventionalcommits.org/) 格式（`feat:` / `fix:` / `refactor:` / `docs:` / `test:` / `chore:`）
- [ ] 分支命名规范：`feature/<描述>` 或 `fix/<描述>`
- [ ] 公共 API 附带 Doxygen 注释
- [ ] 无 `TODO` / `FIXME` / `HACK` 残留标记
- [ ] 无硬编码路径或平台相关代码（跨平台兼容使用 `platform.h` 抽象层）

## 影响范围

<!-- 列出受影响的模块和关键文件 -->

| 模块 | 说明 |
|------|------|
|      |      |

---

> **提示**：CI 会自动运行质量检查和测试。如果 CI 报错，请根据日志修复后重新推送即可。
>
> 如有疑问，欢迎在 PR 中直接提问。


See merge request: openairymax/agentos!242

+++ agentos/cupolas/src/platform/platform.c
@@ -974,17 +974,17 @@ void cupolas_mem_unlock(void *ptr, size_t size)
 
 int32_t cupolas_atomic_load32(volatile int32_t *ptr)
 {
-    return (int32_t)atomic_load_32((volatile _Atomic long *)ptr, memory_order_seq_cst);
+    return (int32_t)atomic_load_32((volatile _Atomic int *)ptr, memory_order_seq_cst);
 }
 
 void cupolas_atomic_store32(volatile int32_t *ptr, int32_t val)
 {
-    atomic_store_32((volatile _Atomic long *)ptr, (long)val, memory_order_seq_cst);
+    atomic_store_32((volatile _Atomic int *)ptr, (int)val, memory_order_seq_cst);
 }
 
 int32_t cupolas_atomic_add32(volatile int32_t *ptr, int32_t delta)
 {
-    return (int32_t)(atomic_fetch_add_32((volatile _Atomic long *)ptr, (long)delta,
+    return (int32_t)(atomic_fetch_add_32((volatile _Atomic int *)ptr, (int)delta,
                                          memory_order_seq_cst) +
                      delta);
 }
@@ -1006,8 +1006,8 @@ int32_t cupolas_atomic_dec32(volatile int32_t *ptr)
 
 bool cupolas_atomic_cas32(volatile int32_t *ptr, int32_t expected, int32_t desired)
 {
-    long exp = (long)expected;
-    return atomic_compare_exchange_strong_32((volatile _Atomic long *)ptr, &exp, (long)desired,
+    int exp = (int)expected;
+    return atomic_compare_exchange_strong_32((volatile _Atomic int *)ptr, &exp, (int)desired,
                                              memory_order_seq_cst, memory_order_seq_cst);
 }
